+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="1.0.0"></a>
+# [1.0.0](https://github.com/chipp972/singleton-module-registry/compare/v0.4.0...v1.0.0) (2018-09-15)
+
+
+### Features
+
+* injectRegistry function to make DI ([c066e58](https://github.com/chipp972/singleton-module-registry/commit/c066e58))
+
+
+### BREAKING CHANGES
+
+* reduced drastically the API
+
+* removed event emitter
+* removed module inter-dependency management
+* added singleton methods to register module and inject registry
+* 100% coverage
+
+
+
 <a name="0.4.0"></a>
 # [0.4.0](https://github.com/chipp972/singleton-module-registry/compare/v0.2.0...v0.4.0) (2018-09-14)